Relocate smoothing functions into tools subdirectory
